refactor(palettes): remove built in palettes to canstants
